% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_severity}
\alias{add_severity}
\title{Add severity}
\usage{
add_severity(
  data_tb,
  severity_args_ls,
  appointments_var_1L_chr = "Appointments",
  date_var_1L_chr = "Date",
  disciplines_chr = c("DE Psychology", "Dietetics", "Psychiatry", "Psychology"),
  end_date_dtm = lubridate::ymd("2024-06-30"),
  provider_var_1L_chr = "ProviderID",
  service_var_1L_chr = "Service",
  severity_var_1L_chr = "Severity",
  tenure_var_1L_chr = "Tenure",
  uid_var_1L_chr = "UID"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{severity_args_ls}{Severity arguments (a list)}

\item{appointments_var_1L_chr}{Appointments variable (a character vector of length one), Default: 'Appointments'}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Date'}

\item{disciplines_chr}{Disciplines (a character vector), Default: c("DE Psychology", "Dietetics", "Psychiatry", "Psychology")}

\item{end_date_dtm}{End date (a date vector), Default: lubridate::ymd("2024-06-30")}

\item{provider_var_1L_chr}{Provider variable (a character vector of length one), Default: 'ProviderID'}

\item{service_var_1L_chr}{Service variable (a character vector of length one), Default: 'Service'}

\item{severity_var_1L_chr}{Severity variable (a character vector of length one), Default: 'Severity'}

\item{tenure_var_1L_chr}{Tenure variable (a character vector of length one), Default: 'Tenure'}

\item{uid_var_1L_chr}{Unique identifier variable (a character vector of length one), Default: 'UID'}
}
\value{
Data (a tibble)
}
\description{
add_severity() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add severity. The function returns Data (a tibble).
}
