% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_sports_data}
\alias{add_sports_data}
\title{Add sports data}
\usage{
add_sports_data(
  datasets_ls,
  categories_chr = c("Risky", "Subjective", "Team", "Type", "Weighed", "Winter"),
  drop_sport_1L_lgl = FALSE
)
}
\arguments{
\item{datasets_ls}{Datasets (a list)}

\item{categories_chr}{Categories (a character vector), Default: c("Risky", "Subjective", "Team", "Type", "Weighed", "Winter")}

\item{drop_sport_1L_lgl}{Drop sport (a logical vector of length one), Default: FALSE}
}
\value{
Datasets (a list)
}
\description{
add_sports_data() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add sports data. The function returns Datasets (a list).
}
\keyword{internal}
