% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athletic-package.R, R/pkg_athletic.R
\docType{package}
\name{athletic-package}
\alias{athletic-package}
\title{athletic: Analyse Health Data for Athletes and Sports Support Staff}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for undertaking analysis of sports datasets for use in computational health economic models developed with the ready4 framework (https://ready4-dev.github.io/ready4/). This development version of the athletic package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton2@monash.edu).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for undertaking analysis of sports datasets for use in computational health economic models developed with the ready4 framework (https://ready4-dev.github.io/ready4/). This development version of the athletic package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton2@monash.edu).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://muheg.github.io/athletic/}
  \item \url{https://github.com/MUHEG/athletic}
  \item \url{https://ready4-dev.github.io/ready4/}
}


Useful links:
\itemize{
  \item \url{https://muheg.github.io/athletic/}
  \item \url{https://github.com/MUHEG/athletic}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Fikru Rizal \email{fikru.rizal@monash.edu} (\href{https://orcid.org/0000-0002-0064-3255}{ORCID})
}

Other contributors:
\itemize{
  \item Monash University [copyright holder]
}

}
\keyword{internal}
