% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Get raw data}
\usage{
get_raw_data(
  path_1L_chr,
  referrals_cols_int = 4L:8L,
  sheets_ls = NULL,
  sheets_int = 1L,
  tabs_chr = character(0)
)
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one)}

\item{referrals_cols_int}{Referrals columns (an integer vector), Default: 4L:8L}

\item{sheets_ls}{Sheets (a list), Default: NULL}

\item{sheets_int}{Sheets (an integer vector), Default: 1}

\item{tabs_chr}{Tabs (a character vector), Default: character(0)}
}
\value{
Datasets (a list)
}
\description{
get_raw_data() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get raw data. The function returns Datasets (a list).
}
\keyword{internal}
