% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_sports_vars}
\alias{get_sports_vars}
\title{Get sports variables}
\usage{
get_sports_vars(
  data_df = NULL,
  exclude_chr = character(0),
  group_1L_chr = character(0)
)
}
\arguments{
\item{data_df}{Data (a data.frame), Default: NULL}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}

\item{group_1L_chr}{Group (a character vector of length one), Default: character(0)}
}
\value{
Sports variables (a character vector)
}
\description{
get_sports_vars() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get sports variables. The function returns Sports variables (a character vector).
}
\keyword{internal}
