% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_betas_plot}
\alias{make_betas_plot}
\title{Make betas plot}
\usage{
make_betas_plot(
  data_ls,
  alpha_1L_dbl = 1,
  bar_width_1L_dbl = 0.25,
  colours_chr = NULL,
  dot_size_1L_dbl = 2.5,
  drop_var_1_chr = NULL,
  drop_var_2_chr = NULL,
  gap_1L_dbl = 0.1,
  grouping_1L_lgl = FALSE,
  hline_xx = ggplot2::geom_hline(yintercept = 0, colour = "grey40"),
  keep_var_1_chr = NULL,
  keep_var_2_chr = NULL,
  labels_chr = NULL,
  legend_name_1L_chr = "Models",
  line_size_1L_dbl = 0.5,
  line_width_1L_dbl = 0,
  plot_type_1L_chr = "dot",
  point_shapes_int = NULL,
  reference_row_1L_lgl = TRUE,
  shade_1L_lgl = FALSE,
  style_1L_chr = "lancet",
  theme_choice_xx = ggplot2::theme_classic(base_family = "Arial Narrow"),
  title_1L_chr = NULL,
  type_1L_chr = "ggsci",
  vertical_1L_lgl = FALSE,
  wrap_width_1L_dbl = 8,
  x_label_1L_chr = "",
  y_label_1L_chr = "Estimates and 95\% CIs"
)
}
\arguments{
\item{data_ls}{Data (a list)}

\item{alpha_1L_dbl}{Alpha (a double vector of length one), Default: 1}

\item{bar_width_1L_dbl}{Bar width (a double vector of length one), Default: 0.25}

\item{colours_chr}{Colours (a character vector), Default: NULL}

\item{dot_size_1L_dbl}{Dot size (a double vector of length one), Default: 2.5}

\item{drop_var_1_chr}{Drop variable 1 (a character vector), Default: NULL}

\item{drop_var_2_chr}{Drop variable 2 (a character vector), Default: NULL}

\item{gap_1L_dbl}{Gap (a double vector of length one), Default: 0.1}

\item{grouping_1L_lgl}{Grouping (a logical vector of length one), Default: FALSE}

\item{hline_xx}{Horizonal line (an output object of multiple potential types), Default: ggplot2::geom_hline(yintercept = 0, colour = "grey40")}

\item{keep_var_1_chr}{Keep variable 1 (a character vector), Default: NULL}

\item{keep_var_2_chr}{Keep variable 2 (a character vector), Default: NULL}

\item{labels_chr}{Labels (a character vector), Default: NULL}

\item{legend_name_1L_chr}{Legend name (a character vector of length one), Default: 'Models'}

\item{line_size_1L_dbl}{Line size (a double vector of length one), Default: 0.5}

\item{line_width_1L_dbl}{Line width (a double vector of length one), Default: 0}

\item{plot_type_1L_chr}{Plot type (a character vector of length one), Default: 'dot'}

\item{point_shapes_int}{Point shapes (an integer vector), Default: NULL}

\item{reference_row_1L_lgl}{Reference row (a logical vector of length one), Default: TRUE}

\item{shade_1L_lgl}{Shade (a logical vector of length one), Default: FALSE}

\item{style_1L_chr}{Style (a character vector of length one), Default: 'lancet'}

\item{theme_choice_xx}{Theme choice (an output object of multiple potential types), Default: ggplot2::theme_classic(base_family = "Arial Narrow")}

\item{title_1L_chr}{Title (a character vector of length one), Default: NULL}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'ggsci'}

\item{vertical_1L_lgl}{Vertical (a logical vector of length one), Default: FALSE}

\item{wrap_width_1L_dbl}{Wrap width (a double vector of length one), Default: 8}

\item{x_label_1L_chr}{X label (a character vector of length one), Default: ''}

\item{y_label_1L_chr}{Y label (a character vector of length one), Default: 'Estimates and 95\% CIs'}
}
\value{
Plot (a plot)
}
\description{
make_betas_plot() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make betas plot. The function returns Plot (a plot).
}
\keyword{internal}
