% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_betas_tbls_ls}
\alias{make_betas_tbls_ls}
\title{Make betas tables list}
\usage{
make_betas_tbls_ls(
  models_ls,
  labels_ls,
  indices_int = NULL,
  element_names_chr = NULL
)
}
\arguments{
\item{models_ls}{Models (a list)}

\item{labels_ls}{Labels (a list)}

\item{indices_int}{Indices (an integer vector), Default: NULL}

\item{element_names_chr}{Element names (a character vector), Default: NULL}
}
\value{
Transformed tables (a list)
}
\description{
make_betas_tbls_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make betas tables list. The function returns Transformed tables (a list).
}
\keyword{internal}
