% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_clinical_vars}
\alias{make_clinical_vars}
\title{Make clinical variables}
\usage{
make_clinical_vars(
  activity_1L_chr = "Activity",
  clinical_team_1L_chr = "Clinical Team",
  clinician_1L_chr = "Clinician",
  clinician_discipline_1L_chr = "Service",
  duration_1L_chr = "Duration",
  exclude_chr = character(0),
  referrers_1L_chr = "Referrer Role",
  severity_1L_chr = "Severity",
  team_disciplines_1L_chr = "Disciplines"
)
}
\arguments{
\item{activity_1L_chr}{Activity (a character vector of length one), Default: 'Activity'}

\item{clinical_team_1L_chr}{Clinical team (a character vector of length one), Default: 'Clinical Team'}

\item{clinician_1L_chr}{Clinician (a character vector of length one), Default: 'Clinician'}

\item{clinician_discipline_1L_chr}{Clinician discipline (a character vector of length one), Default: 'Service'}

\item{duration_1L_chr}{Duration (a character vector of length one), Default: 'Duration'}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}

\item{referrers_1L_chr}{Referrers (a character vector of length one), Default: 'Referrer Role'}

\item{severity_1L_chr}{Severity (a character vector of length one), Default: 'Severity'}

\item{team_disciplines_1L_chr}{Team disciplines (a character vector of length one), Default: 'Disciplines'}
}
\value{
Clinical variables (a character vector)
}
\description{
make_clinical_vars() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make clinical variables. The function returns Clinical variables (a character vector).
}
\keyword{internal}
