% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_continuous_dss_ls}
\alias{make_continuous_dss_ls}
\title{Make continuous datasets list}
\usage{
make_continuous_dss_ls(
  datasets_ls,
  outcomes_chr,
  vars_chr,
  method_1L_chr = "mean",
  labels_ls = NULL,
  tab_spanner_chr = NULL
)
}
\arguments{
\item{datasets_ls}{Datasets (a list)}

\item{outcomes_chr}{Outcomes (a character vector)}

\item{vars_chr}{Variables (a character vector)}

\item{method_1L_chr}{Method (a character vector of length one), Default: 'mean'}

\item{labels_ls}{Labels (a list), Default: NULL}

\item{tab_spanner_chr}{Tab spanner (a character vector), Default: NULL}
}
\value{
Merged tables (a list)
}
\description{
make_continuous_dss_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make continuous datasets list. The function returns Merged tables (a list).
}
\keyword{internal}
