% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_continuous_smry}
\alias{make_continuous_smry}
\title{Make continuous summary}
\usage{
make_continuous_smry(
  data_tb,
  outcome_1L_chr,
  vars_chr,
  method_1L_chr = "mean",
  labels_ls = NULL,
  pval_method_1L_chr = NULL,
  pval_1L_lgl = TRUE
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{outcome_1L_chr}{Outcome (a character vector of length one)}

\item{vars_chr}{Variables (a character vector)}

\item{method_1L_chr}{Method (a character vector of length one), Default: 'mean'}

\item{labels_ls}{Labels (a list), Default: NULL}

\item{pval_method_1L_chr}{P value method (a character vector of length one), Default: NULL}

\item{pval_1L_lgl}{P value (a logical vector of length one), Default: TRUE}
}
\value{
Table (an output object of multiple potential types)
}
\description{
make_continuous_smry() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make continuous summary. The function returns Table (an output object of multiple potential types).
}
\keyword{internal}
