% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_dictionary_lups}
\alias{make_dictionary_lups}
\title{Make dictionary lookup tables}
\usage{
make_dictionary_lups(periods_chr = paste0("Year", 1:3), period_1L_chr = "Year")
}
\arguments{
\item{periods_chr}{Periods (a character vector), Default: paste0("Year", 1:3)}

\item{period_1L_chr}{Period (a character vector of length one), Default: 'Year'}
}
\value{
Dictionary lookup tables (a list)
}
\description{
make_dictionary_lups() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make dictionary lookup tables. The function returns Dictionary lookup tables (a list).
}
\keyword{internal}
