% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_grouped_boxplot_ls}
\alias{make_grouped_boxplot_ls}
\title{Make grouped boxplot list}
\usage{
make_grouped_boxplot_ls(plot_data_ls, outcomes_chr, labels_ls)
}
\arguments{
\item{plot_data_ls}{Plot data (a list)}

\item{outcomes_chr}{Outcomes (a character vector)}

\item{labels_ls}{Labels (a list)}
}
\value{
Plots (a list)
}
\description{
make_grouped_boxplot_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make grouped boxplot list. The function returns Plots (a list).
}
\keyword{internal}
