% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_imputed_dataset}
\alias{make_imputed_dataset}
\title{Make imputed dataset}
\usage{
make_imputed_dataset(X_Ready4useDyad, Y_Ready4useDyad)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{Y_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
Z (A dataset and data dictionary pair.)
}
\description{
make_imputed_dataset() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make imputed dataset. The function is called for its side effects and does not return a value.
}
\keyword{internal}
