% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_keepers}
\alias{make_keepers}
\title{Make keepers}
\usage{
make_keepers(
  names_chr,
  clinical_vars_chr = make_clinical_vars(),
  keep_cdn_1L_chr = c("All", "Personal", "Provider", "Severity", "Sports"),
  severity_1L_int = 8L,
  sports_vars_chr = get_sports_vars()
)
}
\arguments{
\item{names_chr}{Names (a character vector)}

\item{clinical_vars_chr}{Clinical variables (a character vector), Default: make_clinical_vars()}

\item{keep_cdn_1L_chr}{Keep condition (a character vector of length one), Default: c("All", "Personal", "Provider", "Severity", "Sports")}

\item{severity_1L_int}{Severity (an integer vector of length one), Default: 8}

\item{sports_vars_chr}{Sports variables (a character vector), Default: get_sports_vars()}
}
\value{
Keep (a character vector)
}
\description{
make_keepers() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make keepers. The function returns Keep (a character vector).
}
\keyword{internal}
