% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_key_vars}
\alias{make_key_vars}
\title{Make key variables}
\usage{
make_key_vars(
  add_chr = character(0),
  drop_chr = character(0),
  sort_1L_lgl = FALSE
)
}
\arguments{
\item{add_chr}{Add (a character vector), Default: character(0)}

\item{drop_chr}{Drop (a character vector), Default: character(0)}

\item{sort_1L_lgl}{Sort (a logical vector of length one), Default: FALSE}
}
\value{
Key variables (a character vector)
}
\description{
make_key_vars() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make key variables. The function returns Key variables (a character vector).
}
\keyword{internal}
