% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_linked_ds}
\alias{make_linked_ds}
\title{Make linked dataset}
\usage{
make_linked_ds(
  datasets_ls = NULL,
  disciplines_1L_lgl = TRUE,
  end_date_dtm = lubridate::ymd("2024-06-30"),
  exclude_chr = c("Cost", "Duration"),
  imputed_uid_pfx_chr = "UNK",
  keep_all_1L_lgl = FALSE,
  missing_1L_chr = "0",
  path_1L_chr = character(0),
  price_indices_dbl = numeric(0),
  price_ref_1L_int = 1L,
  provider_id_1L_chr = "ProviderID",
  provider_location_1L_chr = "ProviderState",
  referrals_cols_int = 4:7,
  separation_after_dbl = 3,
  sessions_moderate_int = c(4, 15),
  severity_args_ls = NULL,
  severity_var_1L_chr = "Severity",
  sheets_int = 1:5,
  uid_pfx_1L_chr = "CID",
  uid_vars_chr = c("MedlinksID", "AISID"),
  unit_1L_chr = "month",
  var_ctg_chr = character(0),
  what_1L_chr = c("table", "dyad")
)
}
\arguments{
\item{datasets_ls}{Datasets (a list), Default: NULL}

\item{disciplines_1L_lgl}{Disciplines (a logical vector of length one), Default: TRUE}

\item{end_date_dtm}{End date (a date vector), Default: lubridate::ymd("2024-06-30")}

\item{exclude_chr}{Exclude (a character vector), Default: c("Cost", "Duration")}

\item{imputed_uid_pfx_chr}{Imputed unique identifier prefix (a character vector), Default: 'UNK'}

\item{keep_all_1L_lgl}{Keep all (a logical vector of length one), Default: FALSE}

\item{missing_1L_chr}{Missing (a character vector of length one), Default: '0'}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}

\item{price_indices_dbl}{Price indices (a double vector), Default: numeric(0)}

\item{price_ref_1L_int}{Price reference (an integer vector of length one), Default: 1}

\item{provider_id_1L_chr}{Provider identity (a character vector of length one), Default: 'ProviderID'}

\item{provider_location_1L_chr}{Provider location (a character vector of length one), Default: 'ProviderState'}

\item{referrals_cols_int}{Referrals columns (an integer vector), Default: 4:7}

\item{separation_after_dbl}{Separation after (a double vector), Default: 3}

\item{sessions_moderate_int}{Sessions moderate (an integer vector), Default: c(4, 15)}

\item{severity_args_ls}{Severity arguments (a list), Default: NULL}

\item{severity_var_1L_chr}{Severity variable (a character vector of length one), Default: 'Severity'}

\item{sheets_int}{Sheets (an integer vector), Default: 1:5}

\item{uid_pfx_1L_chr}{Unique identifier prefix (a character vector of length one), Default: 'CID'}

\item{uid_vars_chr}{Unique identifier variables (a character vector), Default: c("MedlinksID", "AISID")}

\item{unit_1L_chr}{Unit (a character vector of length one), Default: 'month'}

\item{var_ctg_chr}{Variable category (a character vector), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: c("table", "dyad")}
}
\value{
Data (an output object of multiple potential types)
}
\description{
make_linked_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make linked dataset. The function returns Data (an output object of multiple potential types).
}
