% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdl_outcomes_ds_ls}
\alias{make_mdl_outcomes_ds_ls}
\title{Make model outcomes dataset list}
\usage{
make_mdl_outcomes_ds_ls(datasets_ls, outcomes_chr, predictors_1L_chr)
}
\arguments{
\item{datasets_ls}{Datasets (a list)}

\item{outcomes_chr}{Outcomes (a character vector)}

\item{predictors_1L_chr}{Predictors (a character vector of length one)}
}
\value{
Dataset (a list)
}
\description{
make_mdl_outcomes_ds_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make model outcomes dataset list. The function returns Dataset (a list).
}
\keyword{internal}
