% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdl_predictors_ds_ls}
\alias{make_mdl_predictors_ds_ls}
\title{Make model predictors dataset list}
\usage{
make_mdl_predictors_ds_ls(ds_tb, outcomes_chr, predictors_chr)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{outcomes_chr}{Outcomes (a character vector)}

\item{predictors_chr}{Predictors (a character vector)}
}
\value{
Dataset (a list)
}
\description{
make_mdl_predictors_ds_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make model predictors dataset list. The function returns Dataset (a list).
}
\keyword{internal}
