% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_merged_continuous_smry}
\alias{make_merged_continuous_smry}
\title{Make merged continuous summary}
\usage{
make_merged_continuous_smry(
  datasets_ls,
  outcomes_chr,
  vars_chr,
  method_1L_chr = "mean",
  labels_ls = NULL,
  tab_spanner_chr = NULL,
  pval_1L_lgl = TRUE,
  pval_method_1L_chr = NULL
)
}
\arguments{
\item{datasets_ls}{Datasets (a list)}

\item{outcomes_chr}{Outcomes (a character vector)}

\item{vars_chr}{Variables (a character vector)}

\item{method_1L_chr}{Method (a character vector of length one), Default: 'mean'}

\item{labels_ls}{Labels (a list), Default: NULL}

\item{tab_spanner_chr}{Tab spanner (a character vector), Default: NULL}

\item{pval_1L_lgl}{P value (a logical vector of length one), Default: TRUE}

\item{pval_method_1L_chr}{P value method (a character vector of length one), Default: NULL}
}
\value{
Merged tables (a list)
}
\description{
make_merged_continuous_smry() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make merged continuous summary. The function returns Merged tables (a list).
}
\keyword{internal}
