% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_merged_mdl_tbl}
\alias{make_merged_mdl_tbl}
\title{Make merged model table}
\usage{
make_merged_mdl_tbl(
  models_ls,
  statistics_ls = NULL,
  tab_spanner_chr = NULL,
  labels_ls = NULL
)
}
\arguments{
\item{models_ls}{Models (a list)}

\item{statistics_ls}{Statistics (a list), Default: NULL}

\item{tab_spanner_chr}{Tab spanner (a character vector), Default: NULL}

\item{labels_ls}{Labels (a list), Default: NULL}
}
\value{
merged_table (An object)
}
\description{
make_merged_mdl_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make merged model table. The function is called for its side effects and does not return a value.
}
\keyword{internal}
