% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_modelling_dss}
\alias{make_modelling_dss}
\title{Make modelling datasets}
\usage{
make_modelling_dss(
  data_tb,
  activity_1L_chr = "Activity",
  athlete_roles_chr = c("Athlete", "AlumniAthlete"),
  appointments_var_1L_chr = "Appointments",
  cancellations_var_1L_chr = "Cancellations",
  clinical_team_1L_chr = "Clinical Team",
  clinician_1L_chr = "Clinician",
  clinician_discipline_1L_chr = "Service",
  components_chr = c("Year", "Quarter", "Week"),
  cost_var_1L_chr = "Cost",
  days_1L_chr = "Weekday",
  duration_1L_chr = "Duration",
  exclude_chr = "Group",
  fns_ls = NULL,
  group_1L_chr = character(0),
  index_1L_chr = "Date",
  key_vars_chr = character(0),
  periods_chr = c("sub", "daily", "weekly", "monthly", "quarterly", "yearly"),
  referrals_var_1L_chr = "Referrals",
  referrers_1L_chr = "Referrer Role",
  severity_1L_chr = "Severity",
  team_disciplines_1L_chr = "Disciplines",
  uid_var_1L_chr = "UID"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{activity_1L_chr}{Activity (a character vector of length one), Default: 'Activity'}

\item{athlete_roles_chr}{Athlete roles (a character vector), Default: c("Athlete", "AlumniAthlete")}

\item{appointments_var_1L_chr}{Appointments variable (a character vector of length one), Default: 'Appointments'}

\item{cancellations_var_1L_chr}{Cancellations variable (a character vector of length one), Default: 'Cancellations'}

\item{clinical_team_1L_chr}{Clinical team (a character vector of length one), Default: 'Clinical Team'}

\item{clinician_1L_chr}{Clinician (a character vector of length one), Default: 'Clinician'}

\item{clinician_discipline_1L_chr}{Clinician discipline (a character vector of length one), Default: 'Service'}

\item{components_chr}{Components (a character vector), Default: c("Year", "Quarter", "Week")}

\item{cost_var_1L_chr}{Cost variable (a character vector of length one), Default: 'Cost'}

\item{days_1L_chr}{Days (a character vector of length one), Default: 'Weekday'}

\item{duration_1L_chr}{Duration (a character vector of length one), Default: 'Duration'}

\item{exclude_chr}{Exclude (a character vector), Default: 'Group'}

\item{fns_ls}{Functions (a list), Default: NULL}

\item{group_1L_chr}{Group (a character vector of length one), Default: character(0)}

\item{index_1L_chr}{Index (a character vector of length one), Default: 'Date'}

\item{key_vars_chr}{Key variables (a character vector), Default: character(0)}

\item{periods_chr}{Periods (a character vector), Default: c("sub", "daily", "weekly", "monthly", "quarterly", "yearly")}

\item{referrals_var_1L_chr}{Referrals variable (a character vector of length one), Default: 'Referrals'}

\item{referrers_1L_chr}{Referrers (a character vector of length one), Default: 'Referrer Role'}

\item{severity_1L_chr}{Severity (a character vector of length one), Default: 'Severity'}

\item{team_disciplines_1L_chr}{Team disciplines (a character vector of length one), Default: 'Disciplines'}

\item{uid_var_1L_chr}{Unique identifier variable (a character vector of length one), Default: 'UID'}
}
\value{
dss_lss (An object)
}
\description{
make_modelling_dss() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make modelling datasets. The function is called for its side effects and does not return a value.
}
\keyword{internal}
