% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_nested_boxplot_ls}
\alias{make_nested_boxplot_ls}
\title{Make nested boxplot list}
\usage{
make_nested_boxplot_ls(
  datasets_ls,
  labels_ls = NULL,
  outcomes_ls,
  predictors_chr
)
}
\arguments{
\item{datasets_ls}{Datasets (a list)}

\item{labels_ls}{Labels (a list), Default: NULL}

\item{outcomes_ls}{Outcomes (a list)}

\item{predictors_chr}{Predictors (a character vector)}
}
\value{
Boxplot (a list of lists)
}
\description{
make_nested_boxplot_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make nested boxplot list. The function returns Boxplot (a list of lists).
}
\keyword{internal}
