% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_outcomes_boxplots}
\alias{make_outcomes_boxplots}
\title{Make outcomes boxplots}
\usage{
make_outcomes_boxplots(
  outcomes_chr,
  predictors_chr,
  data_tb,
  labels_ls = NULL,
  palette_1L_chr = "lancet"
)
}
\arguments{
\item{outcomes_chr}{Outcomes (a character vector)}

\item{predictors_chr}{Predictors (a character vector)}

\item{data_tb}{Data (a tibble)}

\item{labels_ls}{Labels (a list), Default: NULL}

\item{palette_1L_chr}{Palette (a character vector of length one), Default: 'lancet'}
}
\value{
Boxplots (a list)
}
\description{
make_outcomes_boxplots() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make outcomes boxplots. The function returns Boxplots (a list).
}
\keyword{internal}
