% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_processed_providers_dss_ls}
\alias{make_processed_providers_dss_ls}
\title{Make processed providers datasets list}
\usage{
make_processed_providers_dss_ls(data_tb)
}
\arguments{
\item{data_tb}{Data (a tibble)}
}
\value{
P all (a tibble)
}
\description{
make_processed_providers_dss_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make processed providers datasets list. The function returns P all (a tibble).
}
\keyword{internal}
