% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_rename_lup}
\alias{make_rename_lup}
\title{Make rename lookup table}
\usage{
make_rename_lup()
}
\value{
Lookup table (Name correspondences lookup table)
}
\description{
make_rename_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make rename lookup table. The function is called for its side effects and does not return a value.
}
\keyword{internal}
