% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_service_use_vars}
\alias{make_service_use_vars}
\title{Make service use variables}
\usage{
make_service_use_vars(
  X_Ready4useDyad,
  active_base_1L_chr = "Active",
  patterns_ls = list(c("[[:space:]]", "")),
  prefix_1L_chr = "Cumulative",
  separation_after_dbl = c(3, 6),
  service_var_1L_chr = "Service",
  tenure_var_1L_chr = "Tenure"
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{active_base_1L_chr}{Active base (a character vector of length one), Default: 'Active'}

\item{patterns_ls}{Patterns (a list), Default: list(c("[\link{:space:}]", ""))}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Cumulative'}

\item{separation_after_dbl}{Separation after (a double vector), Default: c(3, 6)}

\item{service_var_1L_chr}{Service variable (a character vector of length one), Default: 'Service'}

\item{tenure_var_1L_chr}{Tenure variable (a character vector of length one), Default: 'Tenure'}
}
\value{
Service use (a character vector)
}
\description{
make_service_use_vars() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make service use variables. The function returns Service use (a character vector).
}
\keyword{internal}
