% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_severity_args_ls}
\alias{make_severity_args_ls}
\title{Make severity arguments list}
\usage{
make_severity_args_ls(
  disciplines_ls,
  sessions_ls,
  names_chr = character(0),
  severity_var_1L_chr = "Severity"
)
}
\arguments{
\item{disciplines_ls}{Disciplines (a list)}

\item{sessions_ls}{Sessions (a list)}

\item{names_chr}{Names (a character vector), Default: character(0)}

\item{severity_var_1L_chr}{Severity variable (a character vector of length one), Default: 'Severity'}
}
\value{
Severity arguments (a list)
}
\description{
make_severity_args_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make severity arguments list. The function returns Severity arguments (a list).
}
\keyword{internal}
