% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_severity_vars}
\alias{make_severity_vars}
\title{Make severity variables}
\usage{
make_severity_vars(severity_args_ls, severity_var_1L_chr = "Severity")
}
\arguments{
\item{severity_args_ls}{Severity arguments (a list)}

\item{severity_var_1L_chr}{Severity variable (a character vector of length one), Default: 'Severity'}
}
\value{
Severity variables (a character vector)
}
\description{
make_severity_vars() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make severity variables. The function returns Severity variables (a character vector).
}
\keyword{internal}
