% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sngl_predictor_dss_ls}
\alias{make_sngl_predictor_dss_ls}
\title{Make single predictor datasets list}
\usage{
make_sngl_predictor_dss_ls(ds_tb, outcomes_chr, predictors_ls)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{outcomes_chr}{Outcomes (a character vector)}

\item{predictors_ls}{Predictors (a list)}
}
\value{
Dataset (a list)
}
\description{
make_sngl_predictor_dss_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make single predictor datasets list. The function returns Dataset (a list).
}
\keyword{internal}
