% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sports_categories}
\alias{make_sports_categories}
\title{Make sports categories}
\usage{
make_sports_categories(
  type_1L_int = c("current", "deprecated"),
  which_int = integer(0)
)
}
\arguments{
\item{type_1L_int}{Type (an integer vector of length one), Default: c("current", "deprecated")}

\item{which_int}{Which (an integer vector), Default: integer(0)}
}
\value{
Categories (a character vector)
}
\description{
make_sports_categories() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sports categories. The function returns Categories (a character vector).
}
\keyword{internal}
