% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sports_groups}
\alias{make_sports_groups}
\title{Make sports groups}
\usage{
make_sports_groups(
  datasets_ls,
  categories_chr = c("Risky", "Subjective", "Team", "Type", "Weighed", "Winter"),
  simple_1L_lgl = TRUE
)
}
\arguments{
\item{datasets_ls}{Datasets (a list)}

\item{categories_chr}{Categories (a character vector), Default: c("Risky", "Subjective", "Team", "Type", "Weighed", "Winter")}

\item{simple_1L_lgl}{Simple (a logical vector of length one), Default: TRUE}
}
\value{
Grouped (a tibble)
}
\description{
make_sports_groups() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sports groups. The function returns Grouped (a tibble).
}
\keyword{internal}
