% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sports_tb}
\alias{make_sports_tb}
\title{Make sports tibble}
\usage{
make_sports_tb(
  datasets_ls,
  categories_chr = make_sports_categories("deprecated")
)
}
\arguments{
\item{datasets_ls}{Datasets (a list)}

\item{categories_chr}{Categories (a character vector), Default: make_sports_categories("deprecated")}
}
\value{
Sports (a tibble)
}
\description{
make_sports_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sports tibble. The function returns Sports (a tibble).
}
\keyword{internal}
