% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_stacked_mdl_tbl}
\alias{make_stacked_mdl_tbl}
\title{Make stacked model table}
\usage{
make_stacked_mdl_tbl(models_ls, statistics_ls = NULL, labels_ls = NULL)
}
\arguments{
\item{models_ls}{Models (a list)}

\item{statistics_ls}{Statistics (a list), Default: NULL}

\item{labels_ls}{Labels (a list), Default: NULL}
}
\value{
Merged table (an output object of multiple potential types)
}
\description{
make_stacked_mdl_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make stacked model table. The function returns Merged table (an output object of multiple potential types).
}
\keyword{internal}
