% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_tabular_merge}
\alias{make_tabular_merge}
\title{Make tabular merge}
\usage{
make_tabular_merge(
  datasets_ls,
  vars_chr,
  labels_ls = NULL,
  tab_spanner_chr = NULL,
  method_1L_chr = "mean",
  digits_1L_int = 2
)
}
\arguments{
\item{datasets_ls}{Datasets (a list)}

\item{vars_chr}{Variables (a character vector)}

\item{labels_ls}{Labels (a list), Default: NULL}

\item{tab_spanner_chr}{Tab spanner (a character vector), Default: NULL}

\item{method_1L_chr}{Method (a character vector of length one), Default: 'mean'}

\item{digits_1L_int}{Digits (an integer vector of length one), Default: 2}
}
\value{
merged_table (An object)
}
\description{
make_tabular_merge() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make tabular merge. The function is called for its side effects and does not return a value.
}
\keyword{internal}
