% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_tabular_summary}
\alias{make_tabular_summary}
\title{Make tabular summary}
\usage{
make_tabular_summary(
  data_tb,
  vars_chr,
  labels_ls = NULL,
  method_1L_chr = "mean",
  digits_1L_int = 2,
  missing_1L_chr = "no"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{vars_chr}{Variables (a character vector)}

\item{labels_ls}{Labels (a list), Default: NULL}

\item{method_1L_chr}{Method (a character vector of length one), Default: 'mean'}

\item{digits_1L_int}{Digits (an integer vector of length one), Default: 2}

\item{missing_1L_chr}{Missing (a character vector of length one), Default: 'no'}
}
\value{
Table (an output object of multiple potential types)
}
\description{
make_tabular_summary() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make tabular summary. The function returns Table (an output object of multiple potential types).
}
\keyword{internal}
