% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_fake_ages}
\alias{update_fake_ages}
\title{Update fake ages}
\usage{
update_fake_ages(ages_chr)
}
\arguments{
\item{ages_chr}{Ages (a character vector)}
}
\value{
Ages (a character vector)
}
\description{
update_fake_ages() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update fake ages. The function returns Ages (a character vector).
}
\keyword{internal}
