% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_patterns_ls}
\alias{update_patterns_ls}
\title{Update patterns list}
\usage{
update_patterns_ls(
  base_ls = list(c("[[:space:]]", "")),
  prefix_1L_chr = "Cumulative"
)
}
\arguments{
\item{base_ls}{Base (a list), Default: list(c("[\link{:space:}]", ""))}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Cumulative'}
}
\value{
Patterns (a list)
}
\description{
update_patterns_ls() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update patterns list. The function returns Patterns (a list).
}
\keyword{internal}
