% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_counts}
\alias{add_counts}
\title{Add counts}
\usage{
add_counts(datasets_ls, sum_chr = c("Role"), uid_1L_chr = "Client ID")
}
\arguments{
\item{datasets_ls}{Datasets (a list)}

\item{sum_chr}{Sum (a character vector), Default: c("Role")}

\item{uid_1L_chr}{Unique identifier (a character vector of length one), Default: 'Client ID'}
}
\value{
Datasets (a list)
}
\description{
add_counts() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add counts. The function returns Datasets (a list).
}
