% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_sports_data}
\alias{add_sports_data}
\title{Add sports data}
\usage{
add_sports_data(
  datasets_ls,
  categories_chr = make_sports_categories(),
  drop_sport_1L_lgl = FALSE,
  path_1L_chr = character(0),
  sports_1L_int = 81L,
  sports_tab_1L_int = 2L,
  sport_var_1L_chr = "Medlinks Sport categories",
  sum_chr = c("Role"),
  uid_1L_chr = "Client ID"
)
}
\arguments{
\item{datasets_ls}{Datasets (a list)}

\item{categories_chr}{Categories (a character vector), Default: make_sports_categories()}

\item{drop_sport_1L_lgl}{Drop sport (a logical vector of length one), Default: FALSE}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}

\item{sports_1L_int}{Sports (an integer vector of length one), Default: 81}

\item{sports_tab_1L_int}{Sports tab (an integer vector of length one), Default: 2}

\item{sport_var_1L_chr}{Sport variable (a character vector of length one), Default: 'Medlinks Sport categories'}

\item{sum_chr}{Sum (a character vector), Default: c("Role")}

\item{uid_1L_chr}{Unique identifier (a character vector of length one), Default: 'Client ID'}
}
\value{
Datasets (a list)
}
\description{
add_sports_data() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add sports data. The function returns Datasets (a list).
}
\keyword{internal}
