% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Get raw data}
\usage{
get_raw_data(
  path_1L_chr,
  appointments_rows_1L_int = 30,
  referrals_cols_int = 4:8,
  referrals_rows_1L_int = 9,
  cancellations_rows_1L_int = 247,
  retainer_rows_1L_int = 12,
  neuropsychological_rows_1L_int = 1,
  notes_rows_1L_int = 3
)
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one)}

\item{appointments_rows_1L_int}{Appointments rows (an integer vector of length one), Default: 30}

\item{referrals_cols_int}{Referrals columns (an integer vector), Default: 4:8}

\item{referrals_rows_1L_int}{Referrals rows (an integer vector of length one), Default: 9}

\item{cancellations_rows_1L_int}{Cancellations rows (an integer vector of length one), Default: 247}

\item{retainer_rows_1L_int}{Retainer rows (an integer vector of length one), Default: 12}

\item{neuropsychological_rows_1L_int}{Neuropsychological rows (an integer vector of length one), Default: 1}

\item{notes_rows_1L_int}{Notes rows (an integer vector of length one), Default: 3}
}
\value{
Datasets (a list)
}
\description{
get_raw_data() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get raw data. The function returns Datasets (a list).
}
\keyword{internal}
