% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_nested_boxplot_ls}
\alias{make_nested_boxplot_ls}
\title{Make nested boxplot list}
\usage{
make_nested_boxplot_ls(
  dataset_type_1L_chr = "complete",
  datasets_ls,
  outcomes_ls,
  predictors_ls
)
}
\arguments{
\item{dataset_type_1L_chr}{Dataset type (a character vector of length one), Default: 'complete'}

\item{datasets_ls}{Datasets (a list)}

\item{outcomes_ls}{Outcomes (a list)}

\item{predictors_ls}{Predictors (a list)}
}
\value{
Boxplot (a list of lists)
}
\description{
make_nested_boxplot_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make nested boxplot list. The function returns Boxplot (a list of lists).
}
\keyword{internal}
